+++ eva/executor/delete_executor.py
@@ -15,10 +15,17 @@
 from typing import Iterator
 
 import pandas as pd
+from sqlalchemy import and_, or_
 
 from eva.catalog.catalog_manager import CatalogManager
 from eva.catalog.catalog_type import TableType
+from eva.catalog.models.table_catalog import TableCatalogEntry
 from eva.executor.abstract_executor import AbstractExecutor
+from eva.expression.abstract_expression import ExpressionType
+from eva.expression.comparison_expression import ComparisonExpression
+from eva.expression.constant_value_expression import ConstantValueExpression
+from eva.expression.logical_expression import LogicalExpression
+from eva.expression.tuple_value_expression import TupleValueExpression
 from eva.models.storage.batch import Batch
 from eva.plan_nodes.project_plan import ProjectPlan
 from eva.storage.storage_engine import StorageEngine
@@ -32,6 +39,53 @@ class DeleteExecutor(AbstractExecutor):
         self.predicate = node.where_clause
         self.catalog = CatalogManager()
 
+    def predicate_node_to_filter_clause(
+        self, table: TableCatalogEntry, predicate_node: ComparisonExpression
+    ):
+        filter_clause = None
+        left = predicate_node.get_child(0)
+        right = predicate_node.get_child(1)
+
+        if type(left) == TupleValueExpression:
+            column = left.col_name
+            x = table.columns[column]
+        elif type(left) == ConstantValueExpression:
+            value = left.value
+            x = value
+        else:
+            left_filter_clause = self.predicate_node_to_filter_clause(table, left)
+
+        if type(right) == TupleValueExpression:
+            column = right.col_name
+            y = table.columns[column]
+        elif type(right) == ConstantValueExpression:
+            value = right.value
+            y = value
+        else:
+            right_filter_clause = self.predicate_node_to_filter_clause(table, right)
+
+        if type(predicate_node) == LogicalExpression:
+            if predicate_node.etype == ExpressionType.LOGICAL_AND:
+                filter_clause = and_(left_filter_clause, right_filter_clause)
+            elif predicate_node.etype == ExpressionType.LOGICAL_OR:
+                filter_clause = or_(left_filter_clause, right_filter_clause)
+
+        elif type(predicate_node) == ComparisonExpression:
+            if predicate_node.etype == ExpressionType.COMPARE_EQUAL:
+                filter_clause = x == y
+            elif predicate_node.etype == ExpressionType.COMPARE_GREATER:
+                filter_clause = x > y
+            elif predicate_node.etype == ExpressionType.COMPARE_LESSER:
+                filter_clause = x < y
+            elif predicate_node.etype == ExpressionType.COMPARE_GEQ:
+                filter_clause = x >= y
+            elif predicate_node.etype == ExpressionType.COMPARE_LEQ:
+                filter_clause = x <= y
+            elif predicate_node.etype == ExpressionType.COMPARE_NEQ:
+                filter_clause = x != y
+
+        return filter_clause
+
     def exec(self, *args, **kwargs) -> Iterator[Batch]:
         table_catalog = self.node.table_ref.table.table_obj
         storage_engine = StorageEngine.factory(table_catalog)
@@ -40,5 +94,15 @@ class DeleteExecutor(AbstractExecutor):
             table_catalog.table_type == TableType.STRUCTURED_DATA
         ), "DELETE only implemented for structured data"
 
-        storage_engine.delete(table_catalog, self.predicate)
+        table_to_delete_from = storage_engine._try_loading_table_via_reflection(
+            table_catalog.name
+        )
+
+        sqlalchemy_filter_clause = self.predicate_node_to_filter_clause(
+            table_to_delete_from, predicate_node=self.predicate
+        )
+        # verify where clause and convert to sqlalchemy supported filter
+        # https://stackoverflow.com/questions/34026210/where-filter-from-table-object-using-a-dictionary-or-kwargs
+
+        storage_engine.delete(table_catalog, sqlalchemy_filter_clause)
         yield Batch(pd.DataFrame(["Deleted rows"]))
