+++ test/integration_tests/test_delete_executor.py
@@ -124,7 +124,8 @@ class DeleteExecutorTest(unittest.TestCase):
 
     def test_should_delete_tuple_in_table(self):
         delete_query = """DELETE FROM testDeleteOne WHERE
-               id < 20 OR dummyfloat < 2 AND id < 5 AND 20 > id;"""
+               id < 20 OR dummyfloat < 2 AND id < 5 AND 20 > id
+               AND id <= 20 AND id >= 5 OR id != 15 OR id = 15;"""
         batch = execute_query_fetch_all(delete_query)
 
         query = "SELECT * FROM testDeleteOne;"
